+++ app/Http/Controllers/admin/ProductController.php
@@ -19,7 +19,7 @@ class ProductController extends Controller
 {
     private $productService;
     private $categoryService;
-    public function __construct(ProductServiceInterface  $productService ,CategoryServiceInterface $categoryService)
+    public function __construct(ProductServiceInterface  $productService, CategoryServiceInterface $categoryService)
     {
         $this->productService = $productService;
         $this->categoryService = $categoryService;
@@ -31,6 +31,7 @@ class ProductController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Product::class);
         $products = Product::all();
         $categories = Category::get();
         $products = $this->productService->all($request);
@@ -48,13 +49,14 @@ class ProductController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Product::class);
         $products = Product::all();
         $categories = Category::all();
         $params = [
             'categories' => $categories,
             'products' => $products,
         ];
-        return view('admin.product.add',$params);
+        return view('admin.product.add', $params);
     }
 
     /**
@@ -67,15 +69,15 @@ class ProductController extends Controller
     {
         $data = $request->all();
         try {
-        $this->productService->create($data);
-        $notification = [
-            'message' => 'Thêm sản phẩm thành công!',
-            'alert-type' => 'success'
-        ];
-        return redirect()->route('products.index')->with($notification);
+            $this->productService->create($data);
+            $notification = [
+                'message' => 'Thêm sản phẩm thành công!',
+                'alert-type' => 'success'
+            ];
+            return redirect()->route('products.index')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.store.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi xay ra!',
                 'alert-type' => 'error'
@@ -103,9 +105,10 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', Product::class);
         $categories = Category::get();
         $product = $this->productService->find($id);
-        return view('admin.product.edit' , compact('product','categories')) ;
+        return view('admin.product.edit', compact('product', 'categories'));
     }
 
     /**
@@ -118,8 +121,8 @@ class ProductController extends Controller
     public function update(UpdateProductRequest $request, $id)
     {
         $data = $request->all();
-        try{
-            $this->productService->update( $id, $data);
+        try {
+            $this->productService->update($id, $data);
             $notification = [
                 'message' => 'Cập nhật thành công!',
                 'alert-type' => 'success'
@@ -127,7 +130,7 @@ class ProductController extends Controller
             return redirect()->route('products.index')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.update.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
@@ -144,7 +147,8 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        try{
+        $this->authorize('delete', Product::class);
+        try {
             $product = $this->productService->delete($id);
             $notification = [
                 'message' => 'Sản phẩm đã vào thùng rác!',
@@ -153,7 +157,7 @@ class ProductController extends Controller
             return redirect()->route('products.index')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.destroy.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
@@ -162,13 +166,16 @@ class ProductController extends Controller
         }
     }
 
-    public function trash(){
+    public function trash()
+    {
         $products = $this->productService->getTrash();
         return view('admin.product.trash', compact('products'));
     }
 
-    public function restore($id){
-        try{
+    public function restore($id)
+    {
+        $this->authorize('restore', Product::class);
+        try {
             $this->productService->restore($id);
             $notification = [
                 'message' => 'Đã khôi phục sản phẩm!',
@@ -177,7 +184,7 @@ class ProductController extends Controller
             return redirect()->route('product.trash')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.restore.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
@@ -189,8 +196,10 @@ class ProductController extends Controller
         return Excel::download(new ProductsExport, 'products.xlsx');
     }
 
-    public function forcedelete($id){
-        try{
+    public function forcedelete($id)
+    {
+        try {
+            $this->authorize('forceDelete', Product::class);
             $this->productService->forceDelete($id);
             $notification = [
                 'message' => 'Đã xóa sản phẩm',
@@ -199,7 +208,7 @@ class ProductController extends Controller
             return redirect()->route('product.trash')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.forceDelete.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
